* further adjusted setup and dependencies to fully support foreign image loader library like glide * adapter now clear their items correctly for smoother adapter scrolling appearance
