+++ core/src/main/java/com/benny/openlauncher/core/interfaces/App.java
@@ -1,9 +1,11 @@
 package com.benny.openlauncher.core.interfaces;
 
-public interface App extends IconProvider {
+public interface App {
     String getLabel();
 
     String getPackageName();
 
     String getClassName();
+
+    IconProvider getIconProvider();
 }
